Fix huge right margin
The Facebook like button was really wide, such that in mobile the page
had a huge right margin. To fix this I hid the overflow from the content
div. This isn't perfect because it means that the like button text is
cut off on thin screens, but it's better than what was there.